[FLINK-19123] Move PerJobMiniClusterJobClient to top level
