+++ flink-clients/src/main/java/org/apache/flink/client/program/PerJobMiniClusterFactory.java
@@ -18,37 +18,22 @@
 
 package org.apache.flink.client.program;
 
-import org.apache.flink.api.common.JobExecutionResult;
-import org.apache.flink.api.common.JobID;
-import org.apache.flink.api.common.JobStatus;
 import org.apache.flink.configuration.ConfigConstants;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.RestOptions;
 import org.apache.flink.configuration.TaskManagerOptions;
 import org.apache.flink.core.execution.JobClient;
-import org.apache.flink.runtime.concurrent.FutureUtils;
 import org.apache.flink.runtime.jobgraph.JobGraph;
-import org.apache.flink.runtime.jobgraph.OperatorID;
-import org.apache.flink.runtime.jobmaster.JobResult;
 import org.apache.flink.runtime.minicluster.MiniCluster;
 import org.apache.flink.runtime.minicluster.MiniClusterConfiguration;
 import org.apache.flink.runtime.minicluster.RpcServiceSharing;
-import org.apache.flink.runtime.operators.coordination.CoordinationRequest;
-import org.apache.flink.runtime.operators.coordination.CoordinationRequestGateway;
-import org.apache.flink.runtime.operators.coordination.CoordinationResponse;
 import org.apache.flink.util.MathUtils;
-import org.apache.flink.util.SerializedValue;
 import org.apache.flink.util.function.FunctionUtils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.annotation.Nullable;
-
-import java.io.IOException;
-import java.util.Map;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionException;
 import java.util.function.Function;
 
 /**
@@ -139,75 +124,4 @@ public final class PerJobMiniClusterFactory {
 			});
 	}
 
-	/**
-	 * A {@link JobClient} for a {@link PerJobMiniClusterFactory}.
-	 */
-	private static final class PerJobMiniClusterJobClient implements JobClient, CoordinationRequestGateway {
-
-		private final JobID jobID;
-		private final MiniCluster miniCluster;
-		private final CompletableFuture<JobResult> jobResultFuture;
-		private final ClassLoader classLoader;
-
-		private PerJobMiniClusterJobClient(JobID jobID, MiniCluster miniCluster, ClassLoader classLoader) {
-			this.jobID = jobID;
-			this.miniCluster = miniCluster;
-			this.jobResultFuture = miniCluster
-				.requestJobResult(jobID)
-				// Make sure to shutdown the cluster when the job completes.
-				.whenComplete((result, throwable) -> shutDownCluster(miniCluster));
-			this.classLoader = classLoader;
-		}
-
-		@Override
-		public JobID getJobID() {
-			return jobID;
-		}
-
-		@Override
-		public CompletableFuture<JobStatus> getJobStatus() {
-			return miniCluster.getJobStatus(jobID);
-		}
-
-		@Override
-		public CompletableFuture<Void> cancel() {
-			return miniCluster.cancelJob(jobID).thenAccept(result -> {});
-		}
-
-		@Override
-		public CompletableFuture<String> stopWithSavepoint(boolean advanceToEndOfEventTime, @Nullable String savepointDirectory) {
-			return miniCluster.stopWithSavepoint(jobID, savepointDirectory, advanceToEndOfEventTime);
-		}
-
-		@Override
-		public CompletableFuture<String> triggerSavepoint(@Nullable String savepointDirectory) {
-			return miniCluster.triggerSavepoint(jobID, savepointDirectory, false);
-		}
-
-		@Override
-		public CompletableFuture<Map<String, Object>> getAccumulators() {
-			return getJobExecutionResult().thenApply(JobExecutionResult::getAllAccumulatorResults);
-		}
-
-		@Override
-		public CompletableFuture<JobExecutionResult> getJobExecutionResult() {
-			return jobResultFuture.thenApply(result -> {
-				try {
-					return result.toJobExecutionResult(this.classLoader);
-				} catch (Exception e) {
-					throw new CompletionException("Failed to convert JobResult to JobExecutionResult.", e);
-				}
-			});
-		}
-
-		@Override
-		public CompletableFuture<CoordinationResponse> sendCoordinationRequest(OperatorID operatorId, CoordinationRequest request) {
-			try {
-				SerializedValue<CoordinationRequest> serializedRequest = new SerializedValue<>(request);
-				return miniCluster.deliverCoordinationRequestToCoordinator(jobID, operatorId, serializedRequest);
-			} catch (IOException e) {
-				return FutureUtils.completedExceptionally(e);
-			}
-		}
-	}
 }
